+++ jason/src/media/track/remote.rs
@@ -180,7 +180,6 @@ impl Track {
     #[inline]
     pub fn stop(self) {
         if self.0.track.ready_state() == sys::MediaStreamTrackState::Live {
-            self.0.track.stop();
             self.0.on_stopped.call();
         }
     }